requirements.txt
mcp-python-sdk
click
anyio
httpx
